requirements.txt
requests==2.31.0
beautifulsoup4==4.12.2
tweepy==4.14.0
pandas==2.1.0
numpy==1.24.0
openai==1.3.0
transformers==4.35.0
python-dotenv==1.0.0
lxml==4.9.3
pytz==2023.3
